+++ docs/search/all_15.js
@@ -1,7 +1,7 @@
 var searchData=
 [
-  ['webauthentication',['WebAuthentication',['../classBristolSU_1_1Support_1_1Authentication_1_1WebAuthentication.html',1,'BristolSU::Support::Authentication']]],
-  ['webauthentication_2ephp',['WebAuthentication.php',['../WebAuthentication_8php.html',1,'']]],
+  ['websessionauthentication',['WebSessionAuthentication',['../classBristolSU_1_1Support_1_1Authentication_1_1WebSessionAuthentication.html',1,'BristolSU::Support::Authentication']]],
+  ['websessionauthentication_2ephp',['WebSessionAuthentication.php',['../WebSessionAuthentication_8php.html',1,'']]],
   ['withaccesstologicgroupwithresource',['withAccessToLogicGroupWithResource',['../classBristolSU_1_1Support_1_1Logic_1_1Audience_1_1AudienceMemberFactory.html#ac5bf1b5c40f09555bf26d0445d67b501',1,'BristolSU\Support\Logic\Audience\AudienceMemberFactory\withAccessToLogicGroupWithResource()'],['../interfaceBristolSU_1_1Support_1_1Logic_1_1Contracts_1_1Audience_1_1AudienceMemberFactory.html#a8e66b1dde4ff1aca6c72135c12b3578c',1,'BristolSU\Support\Logic\Contracts\Audience\AudienceMemberFactory\withAccessToLogicGroupWithResource()']]],
   ['withaccesstoresource',['withAccessToResource',['../classBristolSU_1_1Support_1_1Logic_1_1Audience_1_1AudienceMemberFactory.html#a70079780995034af7eddee9ccb61741b',1,'BristolSU\Support\Logic\Audience\AudienceMemberFactory\withAccessToResource()'],['../interfaceBristolSU_1_1Support_1_1Logic_1_1Contracts_1_1Audience_1_1AudienceMemberFactory.html#ae26649d898d8ecc3ecb13288fdf49bbb',1,'BristolSU\Support\Logic\Contracts\Audience\AudienceMemberFactory\withAccessToResource()']]]
 ];