+++ docs/search/files_11.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['webauthentication_2ephp',['WebAuthentication.php',['../WebAuthentication_8php.html',1,'']]]
+  ['websessionauthentication_2ephp',['WebSessionAuthentication.php',['../WebSessionAuthentication_8php.html',1,'']]]
 ];