+++ docs/search/files_d.js
@@ -6,6 +6,5 @@ var searchData=
   ['repository_2ephp',['Repository.php',['../Contracts_2Repository_8php.html',1,'(Global Namespace)'],['../Repository_8php.html',1,'(Global Namespace)']]],
   ['resourceidgenerator_2ephp',['ResourceIdGenerator.php',['../ResourceIdGenerator_8php.html',1,'']]],
   ['rolefilter_2ephp',['RoleFilter.php',['../RoleFilter_8php.html',1,'']]],
-  ['rolehasposition_2ephp',['RoleHasPosition.php',['../RoleHasPosition_8php.html',1,'']]],
-  ['roleprovider_2ephp',['RoleProvider.php',['../RoleProvider_8php.html',1,'']]]
+  ['rolehasposition_2ephp',['RoleHasPosition.php',['../RoleHasPosition_8php.html',1,'']]]
 ];