+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1] - (30/01/2020)
+
+### Changed
+- Bumped the control dependency to version 1 (stable)
+
+### Added
+- Laravel Session Authentication
+
+### Removed
+- Laravel Web Authentication
+
+
 ## [1.0.1] - (23/01/2020)
 
 ### Changed
@@ -21,5 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial Release
 
 [Unreleased]: https://github.com/bristol-su/support/compare/v1.0.1...HEAD
-[1.0.1]: https://github.com/bristol-su/support/releases/tag/v1.0...v1.0.1
+[1.1]: https://github.com/bristol-su/support/compare/v1.0.1...v1.1
+[1.0.1]: https://github.com/bristol-su/support/compare/v1.0...v1.0.1
 [1.0]: https://github.com/bristol-su/support/releases/tag/v1.0